+++ src/evolduo_app/music/accompaniment.clj
@@ -1,0 +1,4 @@
+(ns evolduo-app.music.accompaniment)
+
+(def patterns [["picking" "Chord picking"]
+               ["fixed" "Fixed chords"]])