+++ src/evolduo_app/music/midi.clj
@@ -1,0 +1,13 @@
+(ns evolduo-app.music.midi)
+
+;; https://en.wikipedia.org/wiki/General_MIDI
+(def instruments [[4 "Honky-tonk Piano"]
+                  [5 "Electric Piano 1"]
+                  [11 "Music Box"]
+                  [19 "Rock Organ"]
+                  [41 "Violin"]
+                  [49 "String Ensemble 1"]
+                  [51 "Synth Strings 1"]
+                  [82 "Synth Lead 2"]
+
+                  ])