+++ src/evolduo_app/views/playground.clj
@@ -4,7 +4,7 @@
             [clojure.string :as str]))
 
 (defn playground [req & {:keys [abc title] :as track}]
-  (let [{:keys [key mode progression chord tempo notes]} (-> req :params)
+  (let [{:keys [key mode progression chord tempo notes instrument accompaniment]} (-> req :params)
         tempo (if (str/blank? tempo) "110" tempo)]          ;; TODO sanitized params
     (base-view
       req
@@ -33,6 +33,25 @@
            [:div.select
             (comps/progression-select progression)]]]
          [:div.field-label.is-normal
+          [:label.label {:for "tempo"} "Tempo"]]
+         [:div.field.mr-2
+          [:div.control
+           [:input.input {:type "number" :name "tempo" :value tempo :min "40" :max "240"}]]]
+         ]
+        [:div.field.is-horizontal
+         [:div.field-label.is-normal
+          [:label.label {:for "instrument"} "Instrument"]]
+         [:div.field.mr-4
+          [:div.control
+           [:div.select
+            (comps/instrument-select instrument)]]]
+         [:div.field-label.is-normal
+          [:label.label {:for "accompaniment"} "Accompaniment"]]
+         [:div.field.mr-4
+          [:div.control
+           [:div.select
+            (comps/accompaniment-pattern-select accompaniment)]]]
+         [:div.field-label.is-normal
           [:label.label {:for "chord"} "Chord"]]
          [:div.field.mr-4
           [:div.control
@@ -44,17 +63,12 @@
           [:div.control
            [:div.select
             (comps/note-type-select notes)]]]
-         [:div.field-label.is-normal
-          [:label.label {:for "tempo"} "Tempo"]]
-         [:div.field.mr-2
-          [:div.control
-           [:input.input {:type "number" :name "tempo" :value tempo :min "40" :max "240"}]]]
          [:div.control
-          [:input.button.is-link {:type "submit" :value "Try"}]]
-         ]]
+          [:input.button.is-link {:type "submit" :value "Try"}]]] ]
        (when (some? abc)
          [:div
-          (comps/abc-track {:chromosome_id 1 :abc abc } :hide-reaction? true)
+          (comps/abc-track {:chromosome_id 1 :abc abc :instrument instrument :accompaniment accompaniment}
+            :hide-reaction? true)
           #_(comps/abc-track {:chromosome_id 2 :abc abc})])
        ]
       :enable-abc? (some? abc)